switching to using fetch in Subscription, refactoring types
