+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import UserAvatar from './components/UserAvatar'
 import { Subscriptions } from './components/Subscriptions'
 import { Subscription } from './components/Subscription'
 
-export default function Home() {
+export default async function Home() {
   return (
     <>
       <UserAvatar />
